# go-backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.HTTPError:
    properties:
      message:
        example: Error description
        type: string
    type: object
...
  services.CreateOrganizationRequest:
    properties:
      name:
        type: string
    type: object
...
  services.MyOrganizationResponse:
    properties:
      organization:
        type: string
      organization_id:
        type: integer
    type: object
...
  services.OrganizationInfoResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  services.OrganizationTableResponse:
    properties:
      display_name:
        type: string
      id:
        type: integer
      name:
        type: string
      table_type:
        type: string
    type: object
  services.OrganizationUnloadPlaceResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  services.OrganizationUserRequest:
    properties:
      is_primary:
        type: boolean
      required_approval:
        type: boolean
      username:
        type: string
    type: object
  services.OrganizationUserResponse:
    properties:
      first_name:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      last_name:
        type: string
      middle_name:
        type: string
      position:
        type: string
      required_approval:
        type: boolean
      username:
        type: string
    type: object
  services.OrganizationWithUsersResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      user_count:
        type: integer
    type: object
...
  services.UpdateOrganizationTablesRequest:
    properties:
      table_ids:
        items:
          type: integer
        type: array
    type: object
  services.UpdateOrganizationUnloadPlacesRequest:
    properties:
      unload_place_ids:
        items:
          type: integer
        type: array
    type: object
  services.UpdateOrganizationUsersRequest:
    properties:
      users:
        items:
          $ref: '#/definitions/services.OrganizationUserRequest'
        type: array
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: API системы управления пропусками (Бюро пропусков)
  title: Systemburo API
  version: "1.0"
paths:
...
  /get-organization:
    get:
      consumes:
      - application/json
      description: Возвращает организацию, к которой привязан текущий авторизованный
        пользователь
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.MyOrganizationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить организацию текущего пользователя
      tags:
      - organizations
...
  /organizations:
    get:
      consumes:
      - application/json
      description: Возвращает список всех организаций (id, name)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.OrganizationInfoResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить все организации
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Создаёт новую организацию. Требует права buropropuskov
      parameters:
      - description: Данные новой организации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.OrganizationInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Создать организацию
      tags:
      - organizations
  /organizations/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет организацию по ID. Нельзя удалить если есть привязанные
        пользователи. Требует права buropropuskov
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Удалить организацию
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Обновляет название организации по ID. Требует права buropropuskov
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные организации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.OrganizationInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновить организацию
      tags:
      - organizations
  /organizations/{id}/tables:
    get:
      consumes:
      - application/json
      description: Возвращает список таблиц, привязанных к организации
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.OrganizationTableResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить таблицы организации
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Заменяет привязку таблиц к организации. Требует права buropropuskov
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: integer
      - description: Список ID таблиц
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateOrganizationTablesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновить таблицы организации
      tags:
      - organizations
  /organizations/{id}/unload-places:
    get:
      consumes:
      - application/json
      description: Возвращает список мест разгрузки, привязанных к организации
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.OrganizationUnloadPlaceResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить места разгрузки организации
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Заменяет привязку мест разгрузки к организации. Требует права buropropuskov
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: integer
      - description: Список ID мест разгрузки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateOrganizationUnloadPlacesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновить места разгрузки организации
      tags:
      - organizations
  /organizations/{id}/users:
    get:
      consumes:
      - application/json
      description: Возвращает список ответственных пользователей организации
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.OrganizationUserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить пользователей организации
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Заменяет список ответственных пользователей организации (replace-стратегия)
      parameters:
      - description: ID организации
        in: path
        name: id
        required: true
        type: integer
      - description: Список ответственных пользователей
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateOrganizationUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновить пользователей организации
      tags:
      - organizations
  /organizations/with-users:
    get:
      consumes:
      - application/json
      description: Возвращает список организаций с количеством привязанных пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.OrganizationWithUsersResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить организации с количеством пользователей
      tags:
      - organizations
  /organizations/with-users-extended:
    get:
      consumes:
      - application/json
      description: Возвращает организации с количеством пользователей и местами разгрузки
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - BearerAuth: []
      summary: Получить организации с расширенной информацией
      tags:
      - organizations
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"